requirements.txt
Flask==2.3.3
camel-tools>=1.4.0
pytest==7.4.0
python-dotenv==1.0.0
